backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1

# ── Groq LLM ─────────────────────────────────────────────
groq==0.9.0

# ── Local NLP (lightweight, CPU-only) ────────────────────
transformers==4.41.2
torch==2.3.0+cpu
sentence-transformers==3.0.1
spacy==3.7.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# ── Document parsing ──────────────────────────────────────
pdfplumber==0.11.0
python-docx==1.1.2

# ── Data / ML ─────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
networkx==3.3
scikit-learn==1.5.0

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0

# ── Utilities ─────────────────────────────────────────────
httpx==0.27.0
aiofiles==23.2.1
structlog==24.2.0